requirements.txt
google-genai>=1.0.0
pyaudio>=0.2.14
numpy>=1.24.0
keyboard>=0.13.5
pyperclip>=1.8.2
pystray>=0.19.5
pillow>=10.0.0
python-dotenv>=1.0.0
